[runtime] don't force ObjectClone-miss if operand has deprecated map
An object with a deprecated Map which has already been cached in
CloneObjectIC feedback is still a valid Map for fast cloning.

This CL removes the code which forces the Miss path to be taken when a
deprecated map is encountered, avoiding a DCHECK failure and preventing
unnecessary allocation of new fast-clone Maps.

BUG=v8:7611, chromium:867958
R=jkummerow@chromium.org, mvstanton@chrommium.org

Change-Id: I9771b00400fb4dda45a62e874a31d9b50630d847