Update patch set 1
Patch Set 1:

Hi, please take a look at a fix for clusterfuzz regression #5 (that I've seen for this patch).

So, as was mentioned previously, the code is the source of truth for the v8 object model, but it's a bit hard to keep track of what exactly that model is modelling from time to time. I had initially assumed a deprecated Map would be a super bad idea to use (and presumably it would no longer be present in feedback vectors by the time it was deprecated), but on further inspection this is not what happens at all, and it seems that deprecation does not mean that a map is no longer useful/trustworthy, and still accurately describes the layout of a given Object. (Correct me if this is not the case 100% of the time, but that's how it seems from experimentation).

So, this change just removes the "miss-if-deprecated" bit and adds a test.

Patch-set: 1
Reviewer: Gerrit User 1162379 <1162379@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Work-in-progress: false