[cpu-profiler] Add flag to always generate accurate line info.
For optimized code there are two modes. This is controlled by
NeedsSourcePositionsForProfiling(). The problem we currently have is
that this mode must be enabled before the code is optimized, otherwise
the source position information will be low quality, even once the CPU
profiler is turned on.

Also add an implication from --future to this new flag, to gather
perf & memory data from the bots.

Bug: v8:7983
Change-Id: I993fc03df83028529eea365b6d336d9f4c0dcd2b