[turbofan] Collect and use FeedbackCell for inlining/calling.
Change the CALL_IC logic in Ignition to collect the FeedbackCell,
which is unique for a given SharedFunctionInfo within a giving
native context, and thus requires only a single comparison to
guard against cross-context for polymorphic call sites.

Introduce a new CheckClosure operator into TurboFan to handle the
dynamic checks in optimized code. This operator carries the feedback
vector slot from the closure.

On the micro-benchmark initially proposed by Slava (on the tracking bug)
we go from

  console.timeEnd: first, 61.177000
  console.timeEnd: second, 1630.639000
  console.timeEnd: third, 859.022000

to

  console.timeEnd: first, 61.694000
  console.timeEnd: second, 217.466000
  console.timeEnd: third, 102.409000

which corresponds roughly to a ~8x performance boost. On the SixSpeed
test suite the generator-es5 benchmark improves from 165ms to 62ms,
which corresponds to a 2.5x improvement.

Bug: v8:2206, v8:7253
Change-Id: I37227a8a83765eca972dd9680d8551de6af88ed9
Cq-Include-Trybots: luci.chromium.try:linux_chromium_rel_ng