[iwyu] Avoid overwriting source file with same content
The generate-header-include-checks.py script generates the cc files to
check that headers can be included in isolation. It currently always
write to all the cc files. This means that after "gclient runhooks",
all these files have to be recompiled.
This CL fixes that to not touch the file if the content is still the
expected content. This avoids hundreds of recompilations after running
the hooks.

R=machenbach@chromium.org

Bug: v8:7965
Change-Id: I93a0abb936863f6c21ed22ba1b7d900b96725fa3