[build] Prepare MB config for splitting tsan builder
The TSAN builder will be split into builder and tester. This adds the config
for the builder, but keeps the configs for the testers to be removed after
the infra side switch.

This also cleans up redundant concurrent-marking configs, since concurrent
marking is on by default. This also removes a config of a builder that no
longer exists.

TBR=sergiyb@chromium.org
NOTRY=true

Bug: v8:8009
Change-Id: Ic7b146505b5059a2508da94d01d156d5e5e2b997