Fix Array lastIndexOf to call [[HasProperty]] before [[Get]]
Also add more test cases of Array lastIndexOf with proxy, inspired by test262.

In the path for sparse arrays, there is no need to check [[HasProperty]]
before [[Get]]. This is because if elements of the array are changed
within the call of lastIndexOf (by changing the prototype, or adding accessors),
the check by UseSparseVariant will fail and the path will not be taken.
The path for sparse arrays is only taken if the array is not attached with user code.

Bug: v8:7813
Cq-Include-Trybots: luci.v8.try:v8_linux_noi18n_rel_ng
Change-Id: Ifd47149f654e92f56d0a1ed6b3debc93718702be