Update patch set 1
Patch Set 1:

> The size should be a multiple of the OS's page size. V8 should be using the correct sizes already, which are 4K for POSIX, 16K for MIPS ports, and 64K for Windows. You may need to add some PPC-specific behavior here to fix this problem:
> 
> https://cs.chromium.org/chromium/src/v8/src/base/platform/platform-posix.cc?rcl=8c84d9712fe801d937c5e7f92ac2d29e797672d1&l=208
> 
> Note that this only changes the behavior in standalone V8. For the embedding in Chrome you would have to change the page allocator there:
> 
> https://cs.chromium.org/chromium/src/base/allocator/partition_allocator/page_allocator_constants.h?rcl=9a7c6714b07e49ceb227874c53beddb811a4a3e6&l=30

Thanks for looking into this Bill!

The StoreBuffer is of size 16384 (1 << (11 + kPointerSizeLog2))
The AllocatePageSize on ppc64 is 64k.
The CommitPageSize on ppc64 is 64k. 
Since you said that the size should be a multiple of the OS's page size, I think we should increase the StoreBuffer size to accomadate a multiple of 64k on ppc64, right?

Looking at the change in 9d1488e4b0b3d9b2a630c602346f1b8417f2a7c4, the size of StoreBuffer was reduced. 

I am wondering if it is acceptable to bump it up again for only ppc64 little endian like this

 +#if defined(V8_TARGET_ARCH_PPC64) && !defined(V8_OS_AIX) && !defined(V8_TARGET_BIG_ENDIAN)
 +  static const int kStoreBufferSize = 1 << (14 + kPointerSizeLog2);
 +#else
    static const int kStoreBufferSize = 1 << (11 + kPointerSizeLog2);
 +#endif

So that the kStoreBufferSize becomes a multiple of the page size on ppc64.

Patch-set: 1
Reviewer: Gerrit User 1190695 <1190695@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>