[Runtime] No allocation sites for literal in oneshot
No need to create allocation site for literals in Oneshot since
they are executed only once. We dont allocate a valid feedback slot
for literals in oneshot, instead a dummy feedback slot is allocated.
In runtime, allocation site is created based on the value of feedback slot.
We need to update compiler/* to  satisfy unit test eventhough in practice
compiler will not optimize the Oneshot code.

Change-Id: I224b3a30f10361cfe9ff63129b36da8230c5e403