Update patch set 2
Patch Set 2:

(11 comments)

> Patch Set 1:
> 
> (11 comments)
> 
> Personally, I'd rather read 2 lines of source than 20 lines of comment describing that source; but if you find these comments helpful, I don't mind adding them.
> 
> I have a few suggestions for improvements though, see below.

Where are the "2 lines of source" that tell us that 1) We need to change to use a different Macro when we change the return type of the function, 2) the 3rd argument need to match the return type of my function, 3) which variety of the macro we should use based on the return type? I wish there are such 2 lines of source I can read to figure out. Is there somewhere? I cannot find such two lines of source.

Patch-set: 2
Reviewer: Gerrit User 1291594 <1291594@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>