Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

LGTM with a minor comment. And sorry for the delay.

Background, for posterity: when the module namespace pointer was added to PrototypeInfo, the latter's weak_cell existed anyway, so reusing that field for a secondary purpose meant that the new optimization effectively used zero additional memory. Now that the namespace pointer is the only customer requiring this field in the PrototypeInfo, we could try to store it in some other field that is not otherwise used for JSModuleNamespace objects ("prototype_users" would be the first candidate to check, but I or someone else would have to think about this more thoroughly to decide whether it's feasible).

Patch-set: 2
Reviewer: Gerrit User 1115935 <1115935@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1