Update patch set 1
Patch Set 1: Commit-Queue+2

I tried it but it's not a whole lot nicer. It also adds an additional branch before \<non-line-terminator> which is probably the more common case... Not that that matters that much either presumably. I'll leave it as is for now.

Patch-set: 1
Reviewer: Gerrit User 1184114 <1184114@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=+2