[Intl] Pass usage option to icu::Collator
This fixes a bug where V8 didn't pass the usage option to ICU.

The usage option has to be passed to ICU as an unicode extension
value. But, the intl spec forbids the use of 'search' or 'standard' as
an unicode extension value in the language tag. Therefore, we must
filter for this 'search' value before returning the correct value for
collation extension in resolvedOptions.

We only filter the value in resolvedOptions().collation, but don't do
so in resolvedOptions().locale as it more expensive. Not filtering the
locale, doesn't change any behavior. If a locale with collation
extension set to 'search' is passed to a new Intl.Collator
constructor, then we just ignore it as per the spec.

Since we store the usage value in ICU, we no longer have to store this
in the JSCollator object. We can just look it up from ICU when
required. We save one word of memory by doing this.

This patch also fixes std::map and std::set to correctly deal with
const char*.

Cq-Include-Trybots: luci.v8.try:v8_linux_noi18n_rel_ng
Change-Id: Icbd1da4f70393397ef3dee43440c1b121e4a144d
Bug: v8:5751