Update patch set 7
Patch Set 7:

(2 comments)

Overall, I am not too happy with this because now we have some a subtle undocumented invariant and two very large functions that must be kept consistent. I realize the invariant was there before, but before your change it was easier to reason about it (since we were only incrementing the index and not setting it to random values that are subtly dependent on the iterated object).

Patch-set: 7
Reviewer: Gerrit User 1118629 <1118629@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>