Update patch set 2
Patch Set 2:

I dont think this would be the right layer, nor in CompilerDispatcherJob (note, that class is going to be reworked fairly substantially soon). The think kept by the parser is a Utf8CharacterStream [1] (which extracts the data from ExternalStrings) - I think the best place for this scope to live would be alongside that CharacterStream so we know the string is locked while the CharacterStream is alive. Note, I've just landed support for cloning character streams, and that will be what the parallel parse/compile tasks will do to share the stream (they won't directly have access to the ExternalString since that can only be accessed on the main thread), so ideally we could clone the external string locking scope there too to ensure its locked if any of the character streams are still alive.

https://cs.chromium.org/chromium/src/v8/src/parsing/scanner-character-streams.cc?q=Scanner-ch&sq=package:chromium&g=0&l=705

Patch-set: 2
Reviewer: Gerrit User 1002763 <1002763@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>