[intl] add c++ helper function "CreateBoundFunction"
Add a C++ helper function called "CreateBoundFunction" as a static
function to the Intl class, and change all existing builtins involving
the creation of a bound function to use it instead.

Also remove the redundant ContextSlot enums on all Intl objects in favor
of a single BoundFunctionContextSlot enum in the Intl class.

Change-Id: Ibd541f5e79828cd1e2cd573a0e23fc93b348a8ad