[embed handlers] Remove V8_EMBEDDED_BYTECODE_HANDLERS define
Now that the flag has been removed, this definition was just tied to
v8_enable_embedded_builtins, so just remove the define and associated
code.

When bytecode handlers are not embedded, this results in a small memory
increase as the builtins table is increased in size by around 768 words.
On the other hand, the all the custom code for serializing and
deserializing the handlers as well as walking the bytecode handler
dispatch table can be removed.

FIX THIS BEFORE LANDING:
Additionally the bytecode handlers are now not lazily deserialized in
such cases and so they are immediately deserialized at start up time.

ALSO FIX THIS BEFORE LANDING:
Also placeholders are inserted into the builtins table which are never
used when initializing the dispatch table.

Bug: v8:8068
Change-Id: Ie60e196f93d8adb8f1111c56f9a342859ebbd4a6