Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > :-)
> 
> It seems like String#replace and String#split are considered to have side effects because they could be passed an object with `@@split`. Is there a larger fix that could be applied here (maybe checking the type of the argument?) or should I abandon this CL.

That's not the reason at all. If we call with a separator that defines its own @@split, we would perform a side-effect check for that function as well.

All you really need to do is to change the test case, and preferably improve the coverage. E.g. with different separators, including ones that define @@split that has or has no side effects.

Patch-set: 1