Merged: [wasm] Fix worker termination while compiling
If a web worker terminates while a wasm finisher task is scheduled, we
try to cancel that task even though the platform already deleted it.
This results in UBSan failures, ASan failures or crashes.

This CL fixes this by deregistering the foreground task when it is
deleted before being executed.
A layout test for this will be added to chromium in
https://crrev.com/c/1209602.

R=​ahaas@chromium.org

Bug: chromium:875579
Change-Id: Idb2488dde9160e467a4d927d1743a451235d26b3
Originally-reviewed-on: https://chromium-review.googlesource.com/1209344
No-Try: true
No-Presubmit: true
No-Treechecks: true