Update patch set 1
Patch Set 1:

The reason I changed it to v8_target_cpu, was the flag was being set twice in an ia32 build.

For the main targets embedded_builtins was set to false because IA32 was false. For the bytecode_builtins_list_generator it was set to true which resulted in it generating the wrong thing and caused the build failure you're now seeing.

My understanding is that if we're building for IA32, we always want this flag to be false regardless of the host architecture, so I thought v8_target_cpu solved this.

Let's just back my CL while we find the real fix. (It broke the gcov bot as well for which I have another fix: https://chromium-review.googlesource.com/c/v8/v8/+/1219326)

Patch-set: 1
Reviewer: Gerrit User 1204345 <1204345@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>